feat: Introduce SmartQuickAddBar and SmartItemInput components for enhanced product addition
