+++ pcomparator/src/applications/ShoppingLists/Ui/ShoppingListDetails/ShoppingListContainer.tsx
@@ -2,7 +2,7 @@ import { Trans } from "@lingui/react/macro";
 import { useMemo } from "react";
 import type { ShoppingListPayload } from "~/ShoppingLists/Domain/Entities/ShoppingList.entity";
 import { ShoppingListItemCard } from "~/ShoppingLists/Ui/ShoppingListDetails/ShoppingListItemCard";
-import { SmartQuickAddBar } from "~/ShoppingLists/Ui/ShoppingListDetails/SmartQuickAddBar";
+import { SmartQuickAddBar } from "~/ShoppingLists/Ui/ShoppingListDetails/SmartQuickAddBar/SmartQuickAddBar";
 import { useShoppingListActions } from "~/ShoppingLists/Ui/ShoppingListDetails/useShoppingListActions";
 
 interface ShoppingListContainerProps {
